Moves Compositor to playrho/d2/part
